TASK(20260827-110525) done
+ Implemented filtering by priority
+ Added lt, lte, gt, and gte to filter by and with

+++ lib/task.h
@@ -23,7 +23,7 @@ typedef struct task_t {
     char *path;
     char *uuid;
     char *name;
-    size_t priority;
+    long priority;
     task_status status;
     Ht(const char *, bool) tags;
     tasks_t *subtasks;
